Query: Fix for GroupBy Constant with element selector produces invalid SQL
Resolves #11993

Issue: We visit element selector to save mappings for binding later. But in the case of constant key, we create a subquery to inject constant key. This made the mappings for element selector invalid since we cleared projection before pushing down.
Fix is to clear multi-projection after all processing is done.