Query: Clear out projection when GroupBy has element selector as anon type
Issue: When ElementSelector is NewExpression/MemberInitExpression, it could easily bind through mapping and we don't need to preserve projection. In this case when projection is preserved and actual expression is complex, we ended up with multiple projections which we failed to lift during aggregate operation.
Fix: is to clear out projection for NewExpression/MemberInitExpression.
ElementSelector with single property requires aggregate expression over member only for server eval. So with complex aggregate it would do streaming group by.
If elementselector is made complex property then aggregate can be made simple and it still translates.

Resolves #12032